# 8362596: RISC-V: Improve _vectorizedHashCode intrinsic