Hi all,

This pull request contains a backport of commit [4189fcba](https://github.com/openjdk/jdk/commit/4189fcbac40943f3b26c3a01938837b4e4762285) from the [openjdk/jdk](https://git.openjdk.org/jdk) repository.

The commit being backported was authored by Yuri Gaevsky on 27 Jul 2025 and was reviewed by Fei Yang and Feilong Jiang.

Thanks!

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [x] [JDK-8362596](https://bugs.openjdk.org/browse/JDK-8362596) needs maintainer approval

### Issue
 * [JDK-8362596](https://bugs.openjdk.org/browse/JDK-8362596): RISC-V: Improve _vectorizedHashCode intrinsic (**Enhancement** - P4 - Approved)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk25u.git pull/42/head:pull/42` \
`$ git checkout pull/42`

Update a local copy of the PR: \
`$ git checkout pull/42` \
`$ git pull https://git.openjdk.org/jdk25u.git pull/42/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 42`

View PR using the GUI difftool: \
`$ git pr show -t 42`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk25u/pull/42.diff">https://git.openjdk.org/jdk25u/pull/42.diff</a>

</details>
<details><summary>Using Webrev</summary>

[Link to Webrev Comment](https://git.openjdk.org/jdk25u/pull/42#issuecomment-3124871428)
</details>